+++ CHANGELOG.md
@@ -51,6 +51,7 @@ The present file will list all changes made to the project; according to the
 ### Added
 - Support for setting a dropdown value like `status` for assets by using the `id` property (ex: `"status": {"id": 5}`) as it shows in the OpenAPI schema for the High-Level API.
   The previous behavior of setting these properties directly (ex: `"status": 5`) is still supported.
+- New EventLog schema, endpoints, and webhook support in the High-Level API v2.2.
 
 ### Changed